Moved Geolocation services to its own namespace inside CLI module
