+++ module/CLI/config/dependencies.config.php
@@ -35,7 +35,7 @@ return [
             SymfonyCli\Helper\ProcessHelper::class => ProcessHelperFactory::class,
             PhpExecutableFinder::class => InvokableFactory::class,
 
-            Util\GeolocationDbUpdater::class => ConfigAbstractFactory::class,
+            GeoLite\GeolocationDbUpdater::class => ConfigAbstractFactory::class,
             Util\ProcessRunner::class => ConfigAbstractFactory::class,
 
             ApiKey\RoleResolver::class => ConfigAbstractFactory::class,
@@ -70,7 +70,7 @@ return [
     ],
 
     ConfigAbstractFactory::class => [
-        Util\GeolocationDbUpdater::class => [
+        GeoLite\GeolocationDbUpdater::class => [
             DbUpdater::class,
             Reader::class,
             LOCAL_LOCK_FACTORY,
@@ -92,7 +92,7 @@ return [
         Command\ShortUrl\GetShortUrlVisitsCommand::class => [Visit\VisitsStatsHelper::class],
         Command\ShortUrl\DeleteShortUrlCommand::class => [Service\ShortUrl\DeleteShortUrlService::class],
 
-        Command\Visit\DownloadGeoLiteDbCommand::class => [Util\GeolocationDbUpdater::class],
+        Command\Visit\DownloadGeoLiteDbCommand::class => [GeoLite\GeolocationDbUpdater::class],
         Command\Visit\LocateVisitsCommand::class => [
             Visit\VisitLocator::class,
             IpLocationResolverInterface::class,
